+++ apps/api-gateway/src/interfaces/IConnectionSearch.interface.ts
@@ -8,3 +8,18 @@ export interface IConnectionSearchCriteria {
     searchByText: string;
     user?: IUserRequestInterface
 }
+
+export interface IConnectionDetailsById {
+    id: string;
+    createdAt: string;
+    did: string;
+    theirDid: string;
+    theirLabel: string;
+    state: string;
+    role: string;
+    autoAcceptConnection: boolean;
+    threadId: string;
+    protocol: string;
+    outOfBandId: string;
+    updatedAt: string;
+  }
