+++ apps/api-gateway/src/interfaces/ISchemaSearch.interface.ts
@@ -9,18 +9,4 @@ export interface ISchemaSearchPayload {
     searchByText?: string;
     user?: IUserRequestInterface
 }
-export interface IConnectionDetailsById {
-    id: string;
-    createdAt: string;
-    did: string;
-    theirDid: string;
-    theirLabel: string;
-    state: string;
-    role: string;
-    autoAcceptConnection: boolean;
-    threadId: string;
-    protocol: string;
-    outOfBandId: string;
-    updatedAt: string;
-  }
